@@ -1,61 +1,32 @@
 {
     "name": "jorpo/specification",
-    "type": "project",
-    "license": "proprietary",
+    "description": "Simple specification pattern implementation.",
+    "type": "concerto-library",
+    "license": "GNU GPLv3",
     "authors": [
         {
             "name": "John Porter",
             "email": "john@designermonkey.co.uk"
         }
     ],
-    "minimum-stability": "dev",
-    "prefer-stable": true,
-    "require-dev": {
-        "phpunit/phpunit": "^6.5",
-        "squizlabs/php_codesniffer": "^3.2",
-        "theseer/phpdox": "^0.10.1",
-        "phploc/phploc": "^4.0",
-        "phpmd/phpmd": "^2.6"
-    },
-    "autoload-dev": {
-        "psr-4": {
-            "Jorpo\\Specification\\": ["tests"]
-        }
+    "config": {
+        "bin-dir": "bin"
     },
     "require": {
-        "php": "^7.2",
-        "ramsey/uuid": "^3.7",
-        "php-ds/php-ds": "^1.2"
+        "php": "^7.4"
     },
     "autoload": {
         "psr-4": {
             "Jorpo\\Specification\\": ["src"]
         }
     },
-    "config": {
-        "bin-dir": "bin"
+    "require-dev": {
+        "phpunit/phpunit": "^9.1",
+        "phpmetrics/phpmetrics": "^2.6"
     },
-    "scripts": {
-        "test": [
-            "@phpunit"
-        ],
-        "clean": [
-            "@phpcs",
-            "@phpmdstdout"
-        ],
-        "full": [
-            "@phpunit",
-            "@phpcs",
-            "@quietphploc",
-            "@phpmd",
-            "@phpdox"
-        ],
-        "phpunit": "php bin/phpunit",
-        "phpcs": "php bin/phpcs",
-        "phploc": "php bin/phploc --log-xml=doc/phploc.xml --count-tests src",
-        "quietphploc": "php bin/phploc --log-xml=doc/phploc.xml --count-tests --quiet src",
-        "phpmd": "php bin/phpmd src xml cleancode,codesize,controversial,design,naming,unusedcode --reportfile doc/pmd.xml --ignore-violations-on-exit",
-        "phpmdstdout": "php bin/phpmd src xml cleancode,codesize,controversial,design,naming,unusedcode",
-        "phpdox": "php bin/phpdox"
+    "autoload-dev": {
+        "psr-4": {
+            "Jorpo\\Specification\\": ["tests"]
+        }
     }
 }
