@@ -2,30 +2,16 @@
 
 namespace Jorpo\Specification;
 
-/**
- * NOT decorator, used to create a new specifcation that is the inverse (NOT) of the given spec.
- */
 class NotSpecification extends AbstractSpecification
 {
-    /**
-     * @var Specification
-     */
-    private $specification;
+    private Specification $specification;
 
-    /**
-     * Create a new NOT specification based on another spec.
-     *
-     * @param Specification $specification instance to not.
-     */
     public function __construct(Specification $specification)
     {
         $this->specification = $specification;
     }
 
-    /**
-     * @inherit
-     */
-    public function isSatisfiedBy(object $object): bool
+    public function isSatisfiedBy($object): bool
     {
         return !$this->specification->isSatisfiedBy($object);
     }
